requirements.txt
# ETL Pipeline Dependencies
pandas>=1.3.0
requests>=2.25.1
pyyaml>=6.0
python-dotenv>=0.19.0
pytest>=6.2.5
pytest-cov>=2.12.1
black>=21.7b0
isort>=5.9.3
mypy>=0.910
flake8>=3.9.2
apache-airflow>=2.2.0
apache-airflow-providers-postgres>=2.2.0
pyarrow>=6.0.0  # For parquet support